Move client and server logic out of scim-common and into respective modules
* Moved scim-tools-common to scim-tools

NOTE: It's possible that the custom error handling classes that live in common could be replaced with similar classes in scim-spec-protocol & scim-spec-schema
