+++ scim-client/pom.xml
@@ -40,14 +40,6 @@
         <scope>test</scope>
     </dependency>
     <dependency>
-        <groupId>org.apache.directory.scim</groupId>
-        <artifactId>scim-common</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.apache.directory.scim</groupId>
-      <artifactId>scim-common-ws</artifactId>
-    </dependency>
-    <dependency>
       <groupId>org.projectlombok</groupId>
       <artifactId>lombok</artifactId>
       <scope>provided</scope>
