+++ scim-client/src/main/java/org/apache/directory/scim/client/rest/ResourceTypesClient.java
@@ -30,12 +30,6 @@ import jakarta.ws.rs.core.Response;
 
 import org.apache.directory.scim.spec.protocol.ResourceTypesResource;
 import org.apache.directory.scim.spec.schema.ResourceType;
-import org.apache.directory.scim.ws.common.RestClientUtil;
-import org.apache.directory.scim.ws.common.exception.BackingStoreChangedException;
-import org.apache.directory.scim.ws.common.exception.ConflictingDataException;
-import org.apache.directory.scim.ws.common.exception.RestClientException;
-import org.apache.directory.scim.ws.common.exception.RestServerException;
-import org.apache.directory.scim.ws.common.exception.ServiceAuthException;
 
 public class ResourceTypesClient implements AutoCloseable {
