+++ scim-client/src/main/java/org/apache/directory/scim/client/rest/XmlStatusAdapter.java
@@ -1,0 +1,37 @@
+/*
+ * The Pennsylvania State University © 2016
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package org.apache.directory.scim.client.rest;
+
+import jakarta.ws.rs.core.Response.Status;
+import jakarta.xml.bind.annotation.adapters.XmlAdapter;
+
+public final class XmlStatusAdapter extends XmlAdapter<JaxRsStatusAdapterType, Status>
+{
+
+  @Override
+  public JaxRsStatusAdapterType marshal(Status status) throws Exception
+  {
+    return new JaxRsStatusAdapterType(status);
+  }
+
+  @Override
+  public Status unmarshal(JaxRsStatusAdapterType adapterType) throws Exception
+  {
+    return Status.fromStatusCode(adapterType.getCode());
+  }
+
+}
