--- scim-common/src/main/java/org/apache/directory/scim/common/ScimUtils.java
@@ -1,68 +1,0 @@
-/*
-* Licensed to the Apache Software Foundation (ASF) under one
-* or more contributor license agreements.  See the NOTICE file
-* distributed with this work for additional information
-* regarding copyright ownership.  The ASF licenses this file
-* to you under the Apache License, Version 2.0 (the
-* "License"); you may not use this file except in compliance
-* with the License.  You may obtain a copy of the License at
- 
-* http://www.apache.org/licenses/LICENSE-2.0
-
-* Unless required by applicable law or agreed to in writing,
-* software distributed under the License is distributed on an
-* "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-* KIND, either express or implied.  See the License for the
-* specific language governing permissions and limitations
-* under the License.
-*/
-
-package org.apache.directory.scim.common;
-
-import java.lang.reflect.Field;
-import java.text.SimpleDateFormat;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.Date;
-import java.util.List;
-
-public final class ScimUtils {
-  private ScimUtils() {
-  }
-  
-  public static String toDateString(Date date) {
-    SimpleDateFormat dateFormat = new SimpleDateFormat(Constants.ISO_8601_DATE_FORMAT);
-    return dateFormat.format(date);
-  }
-  
-  public static String toDateTimeString(Date date) {
-    SimpleDateFormat dateTimeFormat = new SimpleDateFormat(Constants.ISO_8601_DATE_TIME_FORMAT);
-    return dateTimeFormat.format(date);
-  }
-  
-  public static String toDateString(LocalDate ld) {
-    return ld.format(DateTimeFormatter.ISO_DATE);
-  }
-  
-  public static String toDateTimeString(LocalDateTime ldt) {
-    return ldt.format(DateTimeFormatter.ISO_DATE_TIME);
-  }
-  
-  //Removing the @Nonnull & @Nullable annotations, they are GPL licensed and not compliant with Apache 2.0
-  //public static List<Field> getFieldsUpTo(@Nonnull Class<?> startClass, @Nullable Class<?> exclusiveParent) {
-
-  public static List<Field> getFieldsUpTo(Class<?> startClass, Class<?> exclusiveParent) {
-    List<Field> currentClassFields = new ArrayList<>();
-    Collections.addAll(currentClassFields, startClass.getDeclaredFields());
-    Class<?> parentClass = startClass.getSuperclass();
-    if (parentClass != null && (exclusiveParent == null || !(parentClass.equals(exclusiveParent)))) {
-      List<Field> parentClassFields = getFieldsUpTo(parentClass, exclusiveParent);
-      currentClassFields.addAll(parentClassFields);
-    }
-    return currentClassFields;
-  }
-}
