+++ scim-coverage/pom.xml
@@ -38,10 +38,6 @@
       </dependency>
       <dependency>
         <groupId>org.apache.directory.scim</groupId>
-        <artifactId>scim-common</artifactId>
-      </dependency>
-      <dependency>
-        <groupId>org.apache.directory.scim</groupId>
         <artifactId>scim-tools-common</artifactId>
       </dependency>
       <dependency>
